remove unnecessary queries, move mutations and utils
