+++ src/graphql.d.ts
@@ -29,32 +29,12 @@ declare module '*/DeleteListEntryMutation.graphql' {
   export default defaultDocument
 }
 
-declare module '*/EpisodeCount.graphql' {
-  import { DocumentNode } from 'graphql'
-  const defaultDocument: DocumentNode
-  const EpisodeCount: DocumentNode
-
-  export { EpisodeCount }
-
-  export default defaultDocument
-}
-
 declare module '*/EpisodeList.graphql' {
   import { DocumentNode } from 'graphql'
   const defaultDocument: DocumentNode
   const EpisodeList: DocumentNode
 
   export { EpisodeList }
-
-  export default defaultDocument
-}
-
-declare module '*/ExportQueue.graphql' {
-  import { DocumentNode } from 'graphql'
-  const defaultDocument: DocumentNode
-  const ExportQueue: DocumentNode
-
-  export { ExportQueue }
 
   export default defaultDocument
 }
