--- src/graphql/EpisodeCount.graphql
@@ -1,5 +1,0 @@
-query EpisodeCount($id: Int) {
-  anime: Media(id: $id) {
-    episodes
-  }
-}
