+++ src/graphql/UpdateProgressMutation.graphql
@@ -1,5 +1,5 @@
 mutation UpdateProgressMutation($id: Int, $progress: Int) {
-  SaveMediaListEntry(id: $id, progress: $progress, ) {
+  SaveMediaListEntry(id: $id, progress: $progress) {
     id
     progress
     status
